requirements/base.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.20.0
diffusers>=0.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pillow>=9.0.0
opencv-python-headless>=4.6.0
numpy>=1.21.0
segment-anything>=1.0
omegaconf>=2.3.0
psutil>=5.9.0